pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Josh Haines", email = "Josh@JoshHaines.com"},
]
dependencies = [
    "azure-functions~=1.7",
    "pymongo~=3.11",
]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true
